hps/checkers/checker2/requirements.txt
requests
paho-mqtt
tlv8
pycryptodome==3.15.0